refactor(archives): share extraction state across decoders
