+++ docs/architecture.md
@@ -114,6 +114,23 @@ Deleting Trash streams its own batches, independently of any incomplete measurem
 and GIO URI conversion lives in `adapters/gio_location.rs`, shared by files, operations, preview and
 browser presentation. It preserves native bytes and sanitizes credentials on inbound GIO locations.
 
+Local archive operations live under `adapters/local_operations/archive/`:
+
+| Responsibility | Owner |
+| --- | --- |
+| Operation entry points, worker lifecycle and progress events | `archive.rs` in the parent directory |
+| Staged publication, source traversal and compression writers | `compression.rs` |
+| Per-operation extraction state, copying, cleanup and outcomes | `extraction.rs` |
+| Confined destination writes, path validation and conflict naming | `destination.rs` |
+| ZIP, TAR/gzip and 7z member enumeration, passwords and decoder errors | `decoders.rs` |
+
+Every decoder feeds one `ExtractionSession` per operation. The session has no codec or widget
+API dependencies; decoders lend it member streams and provide already-known remaining locations
+on cancellation. Sequential formats do not scan unread content to complete that list. Format
+libraries and compression behavior are unchanged by this boundary. Decoder evaluation can proceed
+independently of extraction policy. Archive unit tests sit in each module's adjacent `tests.rs`;
+provider-level tests remain in `archive/tests.rs`, with shared test-only builders in `fixtures.rs`.
+
 Feature unit tests sit beside their implementations. Cross-feature browser tests remain in
 `ui/browser/tests/`; GTK tests that need independent initialization can use
 `test_support::gtk_test`, which launches a subprocess with disposable XDG directories. Set
